common/http: Split to reduce dependencies
