+++ oio/common/constants.py
@@ -22,6 +22,9 @@ CHUNK_METADATA_PREFIX = "x-oio-chunk-meta-"
 
 CONTAINER_USER_METADATA_PREFIX = CONTAINER_METADATA_PREFIX + 'user-'
 
+CONNECTION_TIMEOUT = 2.0
+READ_TIMEOUT = 30.0
+
 
 OIO_VERSION = '4.0'
