+++ oio/common/http.py
@@ -13,119 +13,12 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with this library.
 
-
-import logging
 import re
-import socket
-from urllib import quote, quote_plus
 
-from eventlet import patcher
-from eventlet.green.httplib import HTTPConnection, HTTPResponse, _UNKNOWN, \
-        CONTINUE, HTTPMessage
+from urllib import quote_plus
 from oio.common.constants import chunk_headers
-
-urllib3 = patcher.import_patched('urllib3.__init__')
-
-CONNECTION_TIMEOUT = 2.0
-READ_TIMEOUT = 30.0
-
-DEFAULT_POOLSIZE = 32
-DEFAULT_RETRIES = 0
-
-
-class CustomHTTPResponse(HTTPResponse):
-    def __init__(self, sock, debuglevel=0, strict=0,
-                 method=None):
-        self.sock = sock
-        self._actual_socket = sock.fd._sock
-        self.fp = sock.makefile('rb')
-        self.debuglevel = debuglevel
-        self.strict = strict
-        self._method = method
-
-        self.msg = None
-
-        self.version = _UNKNOWN
-        self.status = _UNKNOWN
-        self.reason = _UNKNOWN
-        self.chunked = _UNKNOWN
-        self.chunk_left = _UNKNOWN
-        self.length = _UNKNOWN
-        self.will_close = _UNKNOWN
-
-    def expect_response(self):
-        if self.fp:
-            self.fp.close()
-            self.fp = None
-        self.fp = self.sock.makefile('rb', 0)
-        version, status, reason = self._read_status()
-        if status != CONTINUE:
-            self._read_status = lambda: (version, status, reason)
-            self.begin()
-        else:
-            self.status = status
-            self.reason = reason
-            self.version = 11
-            self.msg = HTTPMessage(self.fp, 0)
-            self.msg.fp = None
-
-    def read(self, amount=None):
-        return HTTPResponse.read(self, amount)
-
-    def force_close(self):
-        if self._actual_socket:
-            self._actual_socket.close()
-        self._actual_socket = None
-        self.close()
-
-    def close(self):
-        HTTPResponse.close(self)
-        self.sock = None
-        self._actual_socket = None
-
-
-class CustomHttpConnection(HTTPConnection):
-    response_class = CustomHTTPResponse
-
-    def connect(self):
-        r = HTTPConnection.connect(self)
-        self.sock.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
-        return r
-
-    def putrequest(self, method, url, skip_host=0, skip_accept_encoding=0):
-        self._method = method
-        self._path = url
-        return HTTPConnection.putrequest(self, method, url, skip_host,
-                                         skip_accept_encoding)
-
-    def getresponse(self):
-        response = HTTPConnection.getresponse(self)
-        logging.debug('HTTP %s %s:%s %s',
-                      self._method, self.host, self.port, self._path)
-        return response
-
-
-def http_connect(host, method, path, headers=None, query_string=None):
-    if isinstance(path, unicode):
-        try:
-            path = path.encode('utf-8')
-        except UnicodeError as e:
-            logging.exception('ERROR encoding to UTF-8: %s', str(e))
-    path = quote('/' + path)
-    conn = CustomHttpConnection(host)
-    if query_string:
-        path += '?' + query_string
-    conn.path = path
-    conn.putrequest(method, path)
-    if headers:
-        for header, value in headers.items():
-            if isinstance(value, list):
-                for k in value:
-                    conn.putheader(header, str(k))
-            else:
-                conn.putheader(header, str(value))
-    conn.endheaders()
-    return conn
+from oio.common.http_eventlet import CustomHttpConnection \
+    as NewCustomHttpConnection
 
 
 _token = r'[^()<>@,;:\"/\[\]?={}\x00-\x20\x7f]+'
@@ -255,23 +148,12 @@ class HeadersDict(dict):
         return dict.pop(self, k.title(), default)
 
 
-def get_pool_manager(pool_connections=DEFAULT_POOLSIZE,
-                     pool_maxsize=DEFAULT_POOLSIZE,
-                     max_retries=DEFAULT_RETRIES):
-    """
-    Get `urllib3.PoolManager` to manage pools of connections
-
-    :param pool_connections: number of connection pools
-    :type pool_connections: `int`
-    :param pool_maxsize: number of connections per connection pool
-    :type pool_maxsize: `int`
-    :param max_retries: number of retries per request
-    :type max_retries: `int`
-    """
-    if max_retries == DEFAULT_RETRIES:
-        max_retries = urllib3.Retry(0, read=False)
-    else:
-        max_retries = urllib3.Retry.from_int(max_retries)
-    return urllib3.PoolManager(num_pools=pool_connections,
-                               maxsize=pool_maxsize, retries=max_retries,
-                               block=False)
+class CustomHttpConnection(NewCustomHttpConnection):
+    def __init__(self, *args, **kwargs):
+        import warnings
+        warnings.simplefilter('once')
+        warnings.warn(
+            "oio.common.http.CustomHttpConnection is deprecated, "
+            "use oio.common.http_eventlet.CustomHttpConnection",
+            DeprecationWarning, stacklevel=2)
+        NewCustomHttpConnection.__init__(self, *args, **kwargs)
