+++ oio/conscience/agent.py
@@ -20,7 +20,7 @@ import pkg_resources
 from eventlet import GreenPool, sleep
 
 from oio.common.daemon import Daemon
-from oio.common.http import get_pool_manager
+from oio.common.http_urllib3 import get_pool_manager
 from oio.common.easy_value import float_value, int_value, true_value
 from oio.common.configuration import parse_config, validate_service_conf
 from oio.common.logger import get_logger
