+++ tests/utils.py
@@ -23,7 +23,7 @@ import testtools
 import random
 import string
 from functools import wraps
-from oio.common.http import get_pool_manager
+from oio.common.http_urllib3 import get_pool_manager
 from urllib import urlencode
 from oio.common.json import json as jsonlib
